+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/MockServerToRbelConverter.java
@@ -6,6 +6,7 @@ package de.gematik.test.tiger.proxy;
 
 import de.gematik.rbellogger.converter.RbelConverter;
 import de.gematik.rbellogger.data.RbelElement;
+import de.gematik.rbellogger.data.RbelElementConvertionPair;
 import de.gematik.rbellogger.data.RbelHostname;
 import de.gematik.rbellogger.data.facet.RbelHttpRequestFacet;
 import de.gematik.rbellogger.data.facet.RbelHttpResponseFacet;
@@ -37,6 +38,7 @@ public class MockServerToRbelConverter {
       HttpResponse response,
       String senderUrl,
       String receiverUrl,
+      CompletableFuture<RbelElement> pairedParsedRequest,
       Optional<ZonedDateTime> timestamp) {
     if (log.isTraceEnabled()) {
       log.trace(
@@ -48,7 +50,7 @@ public class MockServerToRbelConverter {
 
     return rbelConverter
         .parseMessageAsync(
-            responseToRbelMessage(response),
+            new RbelElementConvertionPair(responseToRbelMessage(response), pairedParsedRequest),
             convertUri(senderUrl),
             RbelHostname.fromString(receiverUrl).orElse(null),
             timestamp)
@@ -85,7 +87,7 @@ public class MockServerToRbelConverter {
 
     return rbelConverter
         .parseMessageAsync(
-            requestToRbelMessage(request),
+            new RbelElementConvertionPair(requestToRbelMessage(request)),
             RbelHostname.fromString(request.getRemoteAddress()).orElse(null),
             convertUri(protocolAndHost),
             timestamp)
@@ -134,7 +136,7 @@ public class MockServerToRbelConverter {
 
   private byte[] requestToRawMessage(HttpRequest request) {
     byte[] httpRequestHeader =
-        (request.getMethod().toString()
+        (request.getMethod()
                 + " "
                 + getRequestUrl(request)
                 + " HTTP/1.1\r\n"