+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/tracing/TracingPushController.java
@@ -120,11 +120,13 @@ public class TracingPushController {
 
   private void sendPairedMessage(RbelElement response) {
     final RbelElement request =
-        response.getFacet(TracingMessagePairFacet.class)
+        response
+            .getFacet(TracingMessagePairFacet.class)
             .map(TracingMessagePairFacet::getRequest)
             .or(
                 () ->
-                    response.getFacet(RbelHttpResponseFacet.class)
+                    response
+                        .getFacet(RbelHttpResponseFacet.class)
                         .map(RbelHttpResponseFacet::getRequest))
             .orElseThrow(
                 () ->
@@ -133,13 +135,14 @@ public class TracingPushController {
                             + response.getUuid()
                             + "': Unable to find matching request"));
 
-    RbelTcpIpMessageFacet rbelTcpIpMessageFacet = request.getFacetOrFail(RbelTcpIpMessageFacet.class);
+    RbelTcpIpMessageFacet rbelTcpIpMessageFacet =
+        request.getFacetOrFail(RbelTcpIpMessageFacet.class);
     final RbelHostname sender =
-      RbelHostname.fromString(rbelTcpIpMessageFacet.getSender().getRawStringContent())
-        .orElse(null);
+        RbelHostname.fromString(rbelTcpIpMessageFacet.getSender().getRawStringContent())
+            .orElse(null);
     final RbelHostname receiver =
-      RbelHostname.fromString(rbelTcpIpMessageFacet.getReceiver().getRawStringContent())
-        .orElse(null);
+        RbelHostname.fromString(rbelTcpIpMessageFacet.getReceiver().getRawStringContent())
+            .orElse(null);
     log.trace(
         "{}Propagating new request/response pair (IDs: {} and {})",
         tigerProxy.proxyName(),
@@ -153,7 +156,8 @@ public class TracingPushController {
             .responseUuid(response.getUuid())
             .requestUuid(request.getUuid())
             .responseTransmissionTime(
-                response.getFacet(RbelMessageTimingFacet.class)
+                response
+                    .getFacet(RbelMessageTimingFacet.class)
                     .map(RbelMessageTimingFacet::getTransmissionTime)
                     .orElse(null))
             .requestTransmissionTime(