+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TestTigerProxyExceptionCallbacks.java
@@ -41,17 +41,12 @@ class TestTigerProxyExceptionCallbacks extends AbstractTigerProxyTest {
                         .to("http://localhost:" + fakeBackendServerPort)
                         .build()))
             .build());
-    ReflectionTestUtils.setField(
-        tigerProxy,
-        "mockServerToRbelConverter",
-        new ExceptionThrowingMockRbelConverter(tigerProxy.getRbelLogger().getRbelConverter()));
-    AtomicReference<Throwable> caughtExceptionReference = new AtomicReference<>();
-    tigerProxy.addNewExceptionConsumer(caughtExceptionReference::set);
+    var caughtExceptionReference = setExceptionThrowingMockRbelConverter(tigerProxy);
 
     proxyRest.get("http://backend/foobar").asString();
     await().until(() -> caughtExceptionReference.get() != null);
 
-    assertThat(caughtExceptionReference.get().getMessage()).isEqualTo("foobar");
+    assertThat(caughtExceptionReference.get().getMessage()).contains("foobar");
   }
 
   @Test
@@ -65,19 +60,14 @@ class TestTigerProxyExceptionCallbacks extends AbstractTigerProxyTest {
                         .to("http://localhost:" + fakeBackendServerPort)
                         .build()))
             .build());
-    ReflectionTestUtils.setField(
-        tigerProxy,
-        "mockServerToRbelConverter",
-        new ExceptionThrowingMockRbelConverter(tigerProxy.getRbelLogger().getRbelConverter()));
-    AtomicReference<Throwable> caughtExceptionReference = new AtomicReference<>();
-    tigerProxy.addNewExceptionConsumer(caughtExceptionReference::set);
+    var caughtExceptionReference = setExceptionThrowingMockRbelConverter(tigerProxy);
 
     Unirest.spawnInstance()
         .get("http://localhost:" + tigerProxy.getProxyPort() + "/foobar")
         .asString();
     await().until(() -> caughtExceptionReference.get() != null);
 
-    assertThat(caughtExceptionReference.get().getMessage()).isEqualTo("foobar");
+    assertThat(caughtExceptionReference.get().getMessage()).contains("foobar");
   }
 
   @Test
@@ -91,17 +81,12 @@ class TestTigerProxyExceptionCallbacks extends AbstractTigerProxyTest {
                         .to("http://localhost:" + fakeBackendServerPort)
                         .build()))
             .build());
-    ReflectionTestUtils.setField(
-        tigerProxy,
-        "mockServerToRbelConverter",
-        new ExceptionThrowingMockRbelConverter(tigerProxy.getRbelLogger().getRbelConverter()));
-    AtomicReference<Throwable> caughtExceptionReference = new AtomicReference<>();
-    tigerProxy.addNewExceptionConsumer(caughtExceptionReference::set);
+    var caughtExceptionReference = setExceptionThrowingMockRbelConverter(tigerProxy);
 
     proxyRest.get("http://backend/foobar").asString();
     await().until(() -> caughtExceptionReference.get() != null);
 
-    assertThat(caughtExceptionReference.get().getMessage()).isEqualTo("foobar");
+    assertThat(caughtExceptionReference.get().getMessage()).contains("foobar");
   }
 
   @Test
@@ -115,19 +100,14 @@ class TestTigerProxyExceptionCallbacks extends AbstractTigerProxyTest {
                         .to("http://localhost:" + fakeBackendServerPort)
                         .build()))
             .build());
-    ReflectionTestUtils.setField(
-        tigerProxy,
-        "mockServerToRbelConverter",
-        new ExceptionThrowingMockRbelConverter(tigerProxy.getRbelLogger().getRbelConverter()));
-    AtomicReference<Throwable> caughtExceptionReference = new AtomicReference<>();
-    tigerProxy.addNewExceptionConsumer(caughtExceptionReference::set);
 
+    var caughtExceptionReference = setExceptionThrowingMockRbelConverter(tigerProxy);
     Unirest.spawnInstance()
         .get("http://localhost:" + tigerProxy.getProxyPort() + "/foobar")
         .asString();
     await().until(() -> caughtExceptionReference.get() != null);
 
-    assertThat(caughtExceptionReference.get().getMessage()).isEqualTo("foobar");
+    assertThat(caughtExceptionReference.get().getMessage()).contains("foobar");
   }
 
   public static class ExceptionThrowingMockRbelConverter extends MockServerToRbelConverter {
@@ -141,8 +121,20 @@ class TestTigerProxyExceptionCallbacks extends AbstractTigerProxyTest {
         HttpResponse response,
         String senderUrl,
         String clientAddress,
+        CompletableFuture<RbelElement> pairedParsedRequest,
         Optional<ZonedDateTime> timestamp) {
       throw new RuntimeException("foobar");
     }
+  }
+
+  private AtomicReference<Throwable> setExceptionThrowingMockRbelConverter(TigerProxy tigerProxy) {
+    ReflectionTestUtils.setField(
+        tigerProxy,
+        "mockServerToRbelConverter",
+        new ExceptionThrowingMockRbelConverter(tigerProxy.getRbelLogger().getRbelConverter()));
+    AtomicReference<Throwable> caughtExceptionReference = new AtomicReference<>();
+
+    tigerProxy.addNewExceptionConsumer(caughtExceptionReference::set);
+    return caughtExceptionReference;
   }
 }